requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2016.2.28
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
importlib-metadata==1.6.0
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mock==4.0.2
numpy==1.18.4
protobuf==3.12.1
PyYAML==5.3.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==1.12.2
tensorboardX==2.0
termcolor==1.1.0
threadpoolctl==2.0.0
tqdm==4.46.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
